+++ lib/ui/emulator/input/action/all_actions.dart
@@ -23,6 +23,7 @@ const emulatorActions = [
   togglePause,
   stop,
   toggleFastForward,
+  toggleRewind,
   decreaseVolume,
   increaseVolume,
 ];